refac(front): #470 move medicine recipes
- Move MedicineRecipe to MedicineRecipes/Card
- Move MedicineRecipe/Add to MedicineRecipes/Add
