+++ packages/web/src/views/Landing/MedicineRecipes/Add/translations.ts
