+++ packages/web/src/views/Landing/MedicineRecipes/Card/index.test.tsx
