+++ packages/web/src/views/Landing/MedicineRecipes/Card/index.tsx
