+++ packages/web/src/views/Landing/MedicineRecipes/Card/translations.ts
