+++ packages/web/src/views/Landing/MedicineRecipes/Card/types.ts
